Part 4 - Exercises 4.15 - 4.23*

+++ part4/bloglist/controllers/bloglist.js
@@ -1,22 +1,29 @@
 const blogsRouter = require('express').Router()
+const jwt = require('jsonwebtoken')
+const middleware = require('../utils/middleware')
 const Blog = require('../models/bloglist')
-
+const User = require('../models/user')
+// Exercises 4.17
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
-
-blogsRouter.post('/', async (request, response) => {
+// Exercises 4.19 - 4.22*
+blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
   const body = request.body
+  const user = request.user
 
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
-    likes: body.likes || 0
+    likes: body.likes || 0,
+    user: user._id
   })
 
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save() 
   response.status(201).json(savedBlog)
 })
 
@@ -31,13 +38,18 @@ blogsRouter.put('/:id', async (request, response) => {
   const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {new:true})
   response.json(updatedBlog)
 })
-
-blogsRouter.delete('/:id', async (request, response) => {
-  const blog = await Blog.findByIdAndDelete(request.params.id)
-  if (!blog) {
-    response.status(404).end()
-  }else{
+// Exercises 4.21* - 4.22*
+blogsRouter.delete('/:id', middleware.userExtractor, async (request, response) => {
+  const blog = await Blog.findById(request.params.id)
+  const user = request.user
+  if (!blog){
+    return response.status(404).json({error: 'The Blog not found'})
+  }
+  if (user._id.toString() === blog.user.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
+  } else { 
+    return response.status(401).json({error: 'You are not authorized to delete this blog'})
   }
 })
